requirements.txt
PyMuPDF==1.23.8
azure-storage-blob==12.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
faiss-gpu-cu12
pytesseract
unstructured-client 
unstructured[all-docs]
rank_bm25
streamlit
langchain
langchain_community
langgraph
pynvml
plotly
umap-learn
seaborn
scikit-learn
nbformat